single

+++ runtime/lua/setup.lua
@@ -39,7 +39,7 @@ vim.api.nvim_set_keymap("n", "'", "[external]", {})
 --
 -- settings
 --
-vim.o.ambiwidth = "double"
+vim.o.ambiwidth = "single"
 vim.o.termguicolors = true
 
 vim.o.number = true
